done delete

+++ src/app/home-page/home-page.component.ts
@@ -24,16 +24,24 @@ export class HomePageComponent implements OnInit {
         private route: ActivatedRoute,
         http: HttpClient) {
         http.get<[Product]>('http://5da3dc1aa6593f001407a03e.mockapi.io/api/v1/qlsp').subscribe(res => {
-            this.products = res;
+            this.filteredProduct = res;
         });
         this.config = {
-
             itemsPerPage: 10,
             currentPage: 1,
         };
     }
 
     ngOnInit() {
+        // const id = +this.route.snapshot.paramMap.get('id');
+        // this.productService.getProductById(id).subscribe(() => {
+        //     console.log('success');
+
+        // },
+        //     error => {
+        //         console.log(error);
+        //     });
+
         this.productService.getListProductsByUser().subscribe(
             next => {
                 this.filteredProduct = next;
@@ -42,7 +50,7 @@ export class HomePageComponent implements OnInit {
         );
         this.productService
             .getProducts()
-            .subscribe(next => (this.products = next), error => (this.products = []));
+            .subscribe(next => (this.filteredProduct = next), error => (this.filteredProduct = []));
 
     }
     pageChanged(event) {
@@ -54,17 +62,20 @@ export class HomePageComponent implements OnInit {
         console.log("list " + this.filteredProduct.length);
     }
 
-    //     findProductById(id: number) : Product{
-    //         for(let i =0;i<this.products.l)
-    //         this.products.findIndex
-    // return 
-    //     }
     deletePost(i) {
-        const product = this.products[i];
-        this.productService.deleteProduct(product.id).subscribe(() => {
-            this.products = this.products.filter(t => t.id !== product.id);
-            alert("xoa thanh cong");
-        });
-    }
+        for (let j = 0; j < this.filteredProduct.length - 1; j++) {
+            const product = this.filteredProduct[j];
+            if (product.id === i) {
+                console.log(product);
+                this.productService.deleteProduct(product.id).subscribe(() => {
+                    console.log("delete " + product.id);
+                    const indexOf = this.filteredProduct.indexOf(product);
+                    this.filteredProduct.splice(indexOf, 1);
+                    alert("Delete done");
+                    console.log("Delete");
+                });
+            }
+        }
 
+    }
 }
